timewasted enfpoints

+++ Modules/Mod.cs
@@ -11,7 +11,7 @@
         public bool? track { get; set; } = null;
         public bool? pointFeed { get; set; } = null;
         public double? speed { get; set; } = null;
-        public int? minutesPlayed { get; set; } = null;
+        public float? minutesPlayed { get; set; } = null;
         public bool? aboveAnnounceSpeed { get; set; } = false; 
     }
 }
